requirements.txt
pytest
requests
faker
pydantic
allure-pytest
pytest-xdist
pytest-rerunfailures
jsonschema
python-dotenv
pytest-mock
json-server
flake8>=6.0.0
pre-commit
black
ruff